build: fix library name and SONAME.
CM runtime library is not a module of anything, so fix the name and
the related SONAME to something more meaningful.

Note:
The SONAME is left to libcmrt.so.1 based on the existing ruling from
configure.ac notes. Likewise, library version suffix is made up from
API minor and micro versions.

Signed-off-by: Gwenole Beauchesne <gwenole.beauchesne@intel.com>